Improves a few texts Moves download functions into a new file
